add key_board

+++ handlers.py
@@ -1,8 +1,9 @@
 import logging
 from db import db, find_cafe
-from telegram import InlineQueryResultArticle, InputTextMessageContent
+from telegram import InlineQueryResultArticle, InputTextMessageContent, ReplyKeyboardRemove
 from telegram.ext import ConversationHandler
 from settings import HELP
+from utils import rate_keyboard
 
 CAFE_ICON = 'https://i.ibb.co/pv8RDPL/cafe.png'
 INFO_ICON = 'https://i.ibb.co/4NLgGyD/rating.png'
@@ -48,6 +49,11 @@ def add_or_rate_cafe(update, context):
 
 def add_cafe(update, context):
     text = update.message.text
+    if text == 'Заново':
+        context.bot.send_message(chat_id=update.effective_chat.id,
+                                 text='Напишите название заведения:',
+                                 reply_markup=rate_keyboard())
+        return 'add_cafe_name'
     if 'Cafe_name' in text:
         context.bot.delete_message(chat_id=update.effective_chat.id, message_id=update.message.message_id)
         context.bot.send_message(chat_id=update.effective_chat.id,
@@ -56,150 +62,227 @@ def add_cafe(update, context):
     else:
         context.bot.delete_message(chat_id=update.effective_chat.id, message_id=update.message.message_id)
         text = text.split()
-        context.user_data['cafe_name'] = text[3]
+        context.user_data['cafe_name'] = text[3:]
         cafe_name = context.user_data['cafe_name']
         context.bot.send_message(chat_id=update.effective_chat.id,
                                  parse_mode='HTML',
-                                 text=f'Оцените вкус еды в <b>"{cafe_name}"</b> от 0 до 1.5')
+                                 text=f'Оцените вкус еды в <b>"{cafe_name}"</b> от 0 до 1.5',
+                                 reply_markup=rate_keyboard())
         return 'rate_taste'
+
+
+def add_cafe_name(update, context):
+    if update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    cafe_name = update.message.text
+    context.user_data['cafe_name'] = cafe_name
+    context.bot.send_message(chat_id=update.effective_chat.id,
+                             parse_mode='HTML',
+                             text=f'Оцените вкус еды в <b>"{cafe_name}"</b> от 0 до 1.5')
+    return 'rate_taste'
 
 
 def rate_taste(update, context):
     rating = update.message.text
     cafe_name = context.user_data['cafe_name']
-    try:
-        if 0 <= float(rating) <= 1.5:
-            context.user_data['taste'] = rating
-            update.message.reply_text(f'Оцените подачу в <b>"{cafe_name}"</b> от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_supply'
-        else:
-            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_taste'
-    except ValueError:
-        update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
+    if update.message.text == 'Пропустить':
+        context.user_data['taste'] = 'Без оценки'
+        update.message.reply_text(f'Оцените подачу в <b>"{cafe_name}"</b> от 0 до 1.5',
                                   parse_mode='HTML')
-        return 'rate_taste'
+        return 'rate_supply'
+    elif update.message.text == 'Заново':
+        return add_cafe(update, context)
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        try:
+            if 0 <= float(rating) <= 1.5:
+                context.user_data['taste'] = rating
+                update.message.reply_text(f'Оцените подачу в <b>"{cafe_name}"</b> от 0 до 1.5',
+                                          parse_mode='HTML')
+                return 'rate_supply'
+            else:
+                update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+                return 'rate_taste'
+        except ValueError:
+            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+            return 'rate_taste'
 
 
 def rate_supply(update, context):
     rating = update.message.text
     cafe_name = context.user_data['cafe_name']
-    try:
-        if 0 <= float(rating) <= 1.5:
-            context.user_data['supply'] = rating
-            update.message.reply_text(f'Оцените сервис в <b>"{cafe_name}"</b> от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_service'
-        else:
-            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_supply'
-    except ValueError:
-        update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
+    if update.message.text == 'Пропустить':
+        context.user_data['supply'] = 'Без оценки'
+        update.message.reply_text(f'Оцените сервис в <b>"{cafe_name}"</b> от 0 до 1.5',
                                   parse_mode='HTML')
-        return 'rate_supply'
+        return 'rate_service'
+    elif update.message.text == 'Заново':
+        return add_cafe(update, context)
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        try:
+            if 0 <= float(rating) <= 1.5:
+                context.user_data['supply'] = rating
+                update.message.reply_text(f'Оцените сервис в <b>"{cafe_name}"</b> от 0 до 1.5',
+                                          parse_mode='HTML')
+                return 'rate_service'
+            else:
+                update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+                return 'rate_supply'
+        except ValueError:
+            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+            return 'rate_supply'
 
 
 def rate_service(update, context):
     rating = update.message.text
     cafe_name = context.user_data['cafe_name']
-    try:
-        if 0 <= float(rating) <= 1.5:
-            context.user_data['service'] = rating
-            update.message.reply_text(f'Оцените интерьер в <b>"{cafe_name}"</b> от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_interior'
-        else:
+    if update.message.text == 'Пропустить':
+        context.user_data['service'] = 'Без оценки'
+        update.message.reply_text(f'Оцените интерьер в <b>"{cafe_name}"</b> от 0 до 1.5',
+                                  parse_mode='HTML')
+        return 'rate_interior'
+    elif update.message.text == 'Заново':
+        return add_cafe(update, context)
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        try:
+            if 0 <= float(rating) <= 1.5:
+                context.user_data['service'] = rating
+                update.message.reply_text(f'Оцените интерьер в <b>"{cafe_name}"</b> от 0 до 1.5',
+                                          parse_mode='HTML')
+                return 'rate_interior'
+            else:
+                update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
+                                          parse_mode='HTML')
+                return 'rate_service'
+        except ValueError:
             update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
                                       parse_mode='HTML')
             return 'rate_service'
-    except ValueError:
-        update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
-                                  parse_mode='HTML')
-        return 'rate_service'
 
 
 def rate_interior(update, context):
     rating = update.message.text
     cafe_name = context.user_data['cafe_name']
-    try:
-        if 0 <= float(rating) <= 1.5:
-            context.user_data['interior'] = rating
-            update.message.reply_text(f'Оцените атмосферу в <b>"{cafe_name}"</b> от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_atmosphere'
-        else:
-            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_interior'
-    except ValueError:
-        update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
+    if update.message.text == 'Пропустить':
+        context.user_data['interior'] = 'Без оценки'
+        update.message.reply_text(f'Оцените атмосферу в <b>"{cafe_name}"</b> от 0 до 1.5',
                                   parse_mode='HTML')
-        return 'rate_interior'
+        return 'rate_atmosphere'
+    elif update.message.text == 'Заново':
+        return add_cafe(update, context)
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        try:
+            if 0 <= float(rating) <= 1.5:
+                context.user_data['interior'] = rating
+                update.message.reply_text(f'Оцените атмосферу в <b>"{cafe_name}"</b> от 0 до 1.5',
+                                          parse_mode='HTML')
+                return 'rate_atmosphere'
+            else:
+                update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+                return 'rate_interior'
+        except ValueError:
+            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+            return 'rate_interior'
 
 
 def rate_atmosphere(update, context):
     rating = update.message.text
     cafe_name = context.user_data['cafe_name']
-    try:
-        if 0 <= float(rating) <= 1.5:
-            context.user_data['atmosphere'] = rating
-            update.message.reply_text(f'Оцените маленькие детали в <b>"{cafe_name}"</b> от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_details'
-        else:
-            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
-                                      parse_mode='HTML')
-            return 'rate_atmosphere'
-    except ValueError:
-        update.message.reply_text('Пожалуйста, введите число от 0 до 1.5',
+    if update.message.text == 'Пропустить':
+        context.user_data['atmosphere'] = 'Без оценки'
+        update.message.reply_text(f'Оцените маленькие детали в <b>"{cafe_name}"</b> от 0 до 1.5',
                                   parse_mode='HTML')
-        return 'rate_atmosphere'
+        return 'rate_details'
+    elif update.message.text == 'Заново':
+        return add_cafe(update, context)
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        try:
+            if 0 <= float(rating) <= 1.5:
+                context.user_data['atmosphere'] = rating
+                update.message.reply_text(f'Оцените маленькие детали в <b>"{cafe_name}"</b> от 0 до 1.5',
+                                          parse_mode='HTML')
+                return 'rate_details'
+            else:
+                update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+                return 'rate_atmosphere'
+        except ValueError:
+            update.message.reply_text('Пожалуйста, введите число от 0 до 1.5')
+            return 'rate_atmosphere'
 
 
 def rate_details(update, context):
     rating = update.message.text
     cafe_name = context.user_data['cafe_name']
-    try:
-        if 0 <= float(rating) <= 1.5:
-            context.user_data['details'] = rating
-            update.message.reply_text(f'Добавьте дополнительный балл для <b>"{cafe_name}"</b> от 0 до 1 по желанию',
-                                      parse_mode='HTML')
-            return 'add_point'
-        else:
-            update.message.reply_text('Пожалуйста, введите число от 0 до 1',
-                                      parse_mode='HTML')
-            return 'rate_details'
-    except ValueError:
-        update.message.reply_text('Пожалуйста, введите число от 0 до 1',
+    if update.message.text == 'Пропустить':
+        context.user_data['details'] = 'Без оценки'
+        update.message.reply_text(f'Добавьте дополнительный балл для <b>"{cafe_name}"</b> от 0 до 1 по желанию',
                                   parse_mode='HTML')
-        return 'rate_details'
+        return 'add_point'
+    elif update.message.text == 'Заново':
+        return add_cafe(update, context)
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        try:
+            if 0 <= float(rating) <= 1.5:
+                context.user_data['details'] = rating
+                update.message.reply_text(f'Добавьте дополнительный балл для <b>"{cafe_name}"</b> '
+                                          'от 0 до 1 по желанию',
+                                          parse_mode='HTML')
+                return 'add_point'
+            else:
+                update.message.reply_text('Пожалуйста, введите число от 0 до 1')
+                return 'rate_details'
+        except ValueError:
+            update.message.reply_text('Пожалуйста, введите число от 0 до 1')
+            return 'rate_details'
 
 
 def add_point(update, context):
     point = update.message.text
     cafe_name = context.user_data['cafe_name']
-    try:
-        if 0 <= float(point) <= 1:
-            context.user_data['point'] = point
-            update.message.reply_text(f'Добавьте комментарий для <b>"{cafe_name}"</b>',
-                                      parse_mode='HTML')
-            return 'add_comment'
-        else:
-            update.message.reply_text('Пожалуйста, введите число от 0 до 1',
-                                      parse_mode='HTML')
-            return 'rate_point'
-    except ValueError:
-        update.message.reply_text('Пожалуйста, введите число от 0 до 1',
+    if update.message.text == 'Пропустить':
+        context.user_data['point'] = 'Без оценки'
+        update.message.reply_text(f'Добавьте комментарий для <b>"{cafe_name}"</b>',
                                   parse_mode='HTML')
-        return 'rate_point'
+        return 'add_comment'
+    elif update.message.text == 'Заново':
+        return add_cafe(update, context)
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        try:
+            if 0 <= float(point) <= 1:
+                context.user_data['point'] = point
+                update.message.reply_text(f'Добавьте комментарий для <b>"{cafe_name}"</b>',
+                                          parse_mode='HTML')
+                return 'add_comment'
+            else:
+                update.message.reply_text('Пожалуйста, введите число от 0 до 1')
+                return 'rate_point'
+        except ValueError:
+            update.message.reply_text('Пожалуйста, введите число от 0 до 1')
+            return 'rate_point'
 
 
 def add_comment(update, context):
-    comment = update.message.text
-    context.user_data['comment'] = comment
+    if update.message.text == 'Пропустить':
+        context.user_data['comment'] = 'Без комментариев'
+        comment = context.user_data['comment']
+    elif update.message.text == 'Выйти':
+        return exit_rating(update, context)
+    else:
+        comment = update.message.text
+        context.user_data['comment'] = comment
     cafe_name = context.user_data['cafe_name']
     taste = context.user_data['taste']
     supply = context.user_data['supply']
@@ -208,8 +291,16 @@ def add_comment(update, context):
     atmosphere = context.user_data['atmosphere']
     details = context.user_data['details']
     point = context.user_data['point']
-    summ = float(taste)+float(supply)+float(service)+float(interior)+float(atmosphere)+float(details)+float(point)
-    print(context.user_data)
+    summ = 0
+    values = context.user_data.values()
+    list_values = list(values)
+    for rating in list_values[1:8]:
+        try:
+            summ = summ + float(rating)
+        except ValueError:
+            continue
+    if summ == 0:
+        summ = 'Без оценки'
     update.message.reply_text(f'Название: <b>{cafe_name}</b>\n'
                               f'Вкус - <b>{taste}</b>\n'
                               f'Подача - <b>{supply}</b>\n'
@@ -221,6 +312,14 @@ def add_comment(update, context):
                               f'Комментарий: <b>{comment}</b>\n\n'
                               f'<b>Итого: {summ}</b>',
                               parse_mode='HTML')
+    context.user_data.clear()
+    return ConversationHandler.END
+
+
+def exit_rating(update, context):
+    update.message.reply_text('Оценка удалена. До свидания!',
+                              reply_markup=ReplyKeyboardRemove())
+    context.user_data.clear()
     return ConversationHandler.END
